Aden protectorate formation modified
References issues #59 and #36

Historical flavor is added to portray the events leading to formation of the Aden Protectorate, with some minimum flavor included. It starts with flavor that can lead to the historical occupation of Aden by the United Kingdom. From there it describes the periodic Fadhli raids on Aden, and covers the 1866 British military retribution against Fadhli. An event is added that describes that Kathiri-Qu'aiti conflict that was recorded starting in 1858. For Mahra, the possibility of a Socotra purchase is added, by a Great Power owner of Bombay from certain continents, expected to be the United Kingdom. From there, if conditions are right, the Aden Protectorate will be formed. Informal protection is offered for Kathiri and Mahra before the Aden Protectorate is formed. Player protection exists for most of these stages, but excessively slow westernizers can lose. New graphics are added for events and decisions. National Unification of Yemen no longer removes cores, because some internal instability is desired in Yemen. Other events and decisions affecting the above are commented out.

Minor fixes and tuning for Fadhli

Added an event for Fadhli Punishment, because it seemed wrong to see the war start without explanation. Also, allowed it to occur even when Fadhli is in the sphere of the owner of Aden, because this seems allowable in historical terms. Some localisation was adjusted and added to make the text more flavorful.